Move read bytes utilities to Encoding
